revert some changes for compatibility
