+++ AlembicImporter/Assets/UTJ/Alembic/Scripts/Timeline/AlembicRecorderTrack.cs.meta
@@ -1,7 +1,9 @@
 fileFormatVersion: 2
-guid: 44d84dc3cdf9d044c9bd290c18ab5454
-timeCreated: 1489425277
+guid: 6358aae859449c84db906e350df8e4ea
+timeCreated: 1519256061
+licenseType: Pro
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
