+++ AlembicImporter/Assets/UTJ/Alembic/Scripts/Timeline/AlembicShotPlayable.cs.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
 guid: 9cd9987895ad98e4889ea378e1b047f6
 timeCreated: 1492612433
-licenseType: Free
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
