+++ AlembicImporter/Assets/UTJ/Alembic/Scripts/Timeline/AlembicTrack.cs.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
 guid: f669ed87322f9554d9dff5c29f48f549
 timeCreated: 1489425277
-licenseType: Free
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
